requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.4
django-crispy-forms==1.9.0
django-rest-swagger==2.2.0
django-tables2==2.3.1
djangorestframework==3.11.0
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.8.4
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
selenium==3.141.0
simplejson==3.17.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.18
django-push-notifications
django-filter==2.2.0
django-bootstrap4
django-bootstrap-datepicker-plus
django-phonenumber-field
phonenumbers